switched to ints for line number tracking throughout
Was using a mix of size_t and int before

+++ include/cinder/gl/ShaderPreprocessor.h
@@ -89,11 +89,11 @@ class ShaderPreprocessor {
 	SignalIncludeHandler& getSignalInclude()	{ return mSignalInclude; }
 	
   private:
-	void			parseDirectives( const std::string &source, const fs::path &sourcePath, std::string *directives, std::string *sourceBody, int *versionNumber, size_t *lineNumberStart );
-	std::string		parseTopLevel( const std::string &source, const fs::path &currentDirectory, size_t lineNumberStart, int versionNumber, std::set<fs::path> &includeTree );
+	void			parseDirectives( const std::string &source, const fs::path &sourcePath, std::string *directives, std::string *sourceBody, int *versionNumber, int *lineNumberStart );
+	std::string		parseTopLevel( const std::string &source, const fs::path &currentDirectory, int lineNumberStart, int versionNumber, std::set<fs::path> &includeTree );
 	std::string		parseRecursive( const fs::path &path, const fs::path &currentDirectory, int versionNumber, std::set<fs::path> &includeTree );
 	std::string		readStream( std::istream &stream, const fs::path &path, int versionNumber, std::set<fs::path> &includeTree );
-	std::string		getLineDirective( const fs::path &sourcePath, size_t lineNumber, size_t sourceStringNumber, int versionNumber ) const;
+	std::string		getLineDirective( const fs::path &sourcePath, int lineNumber, int sourceStringNumber, int versionNumber ) const;
 	fs::path		findFullPath( const fs::path &includePath, const fs::path &currentPath );
 	
 	int								mVersion;
